pyproject.toml
[project]
name = "pixelcraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.9",
    "python-decouple>=3.8",
]
